requirements.txt
pandas>=1.5.0
scikit-learn>=1.2.0
mlflow>=2.0.0
dvc>=3.0.0
numpy>=1.23.0
pyyaml>=6.0
matplotlib>=3.5.0
seaborn>=0.12.0
